+++ vm_label.c
@@ -7,10 +7,13 @@
 #define STR_HELPER(x) #x
 #define STR(x) STR_HELPER(x)
 
-#define MAX_LABEL_DIGITS 10
-#define LABEL_FORMAT "L%0" STR( MAX_LABEL_DIGITS ) "d"
+#define CAT3_HELPER( x, y, z ) x ## y ## z 
+#define CAT3( x, y, z)  CAT3_HELPER( x, y, z )
 
-static int vm_label_counter = 0;
+#define MAX_LABEL_DIGITS 10
+#define LABEL_FORMAT CAT3( L%0 , MAX_LABEL_DIGITS , u )
+
+static unsigned int vm_label_counter = 0;
 static char* cur_vm_label;
 
 char* new_vm_label(){
@@ -19,9 +22,12 @@ char* new_vm_label(){
 	vm_label_counter = vm_label_counter + 1 ; // increment label counter.
 
 	if(vm_label_counter <= (pow(10, MAX_LABEL_DIGITS ) - 1)){
-        char* label_format = LABEL_FORMAT ;
-		// printf("%s \n", label_format );
-		sprintf(new_label, label_format , vm_label_counter); 
+        char* label_format = STR( LABEL_FORMAT ) ;
+//		printf("---------LABEL FORMAT-----------\n");
+//		printf("%s\n", label_format);
+//		printf("%s \n", label_format );
+		snprintf(new_label, (MAX_LABEL_DIGITS + 2) * sizeof(char) , label_format, vm_label_counter); 
+//		printf("%s\n", new_label);
 		// No need to call new_label[ MAX_LABEL_DIGITS + 1] = '\0';
 	}else{
 		printf("Compiler exhausted counts for label creation. \n");